updates test cases and made indexedDB in sync with heap
