+++ lib/index.js
@@ -12,8 +12,11 @@ module.exports = {
             idb: require('./indexed-db/index'),
             heap: require('./heap/index')
         };
-        var store = storeSelector && storeSelector();
-        if (dataStoreMap[store]) {
+        if (typeof storeSelector === 'string') {
+            return dataStorePriority[storeSelector];
+        }
+        if (typeof storeSelector === 'function') {
+            var store = storeSelector();
             return dataStoreMap[store];
         }
         for (var i = 0, j = dataStorePriority.length; i < j; i = i + 1) {
