+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "localdatastore",
-    "version": "0.1.7",
+    "version": "0.2.0",
     "description": "Promise based client/server side storage. Uses IndexedDB/heap based on availability",
     "main": "lib/index.js",
     "scripts": {
