+++ test/indexed-db/init.js
@@ -7,27 +7,34 @@ describe('data-store/indexed-db/init', function() {
     }
 
     it('creates a database without errors', function() {
+        var database = null;
         return idb.init('db-init-1', [
             {
-                indexes: [{name: 'attr1'}],
+                indexes: [
+                    {
+                        name: 'attr1'
+                    }
+                ],
                 name: 'table1'
             }
-        ]).then(function() {
-            return idb.insert({name: 'table1'}, [
+        ]).then(function(db) {
+            database = db;
+        }).then(function() {
+            return database.insert({name: 'table1'}, [
                 {
                     attr1: 'A',
                     attr2: 'B'
                 }
-            ]).then(function() {
-                return idb.select({
-                    index: 'attr1',
-                    name: 'table1'
-                }, {}).then(function(data) {
-                    expect(data[0].attr1).to.eql('A');
-                    expect(data[0].attr2).to.eql('B');
-                    return idb.destroy('db-init-1');
-                });
-            });
+            ]);
+        }).then(function() {
+            return database.select({
+                index: 'attr1',
+                name: 'table1'
+            }, {});
+        }).then(function(data) {
+            expect(data[0].attr1).to.eql('A');
+            expect(data[0].attr2).to.eql('B');
+            return idb.destroy('db-init-1');
         });
     });
 });
